Partial fix for broken MSVC build (still need to fix prebuild)
The MSVC2010/ dir is omitted, since it doesn't work for me yet, and adding a
whole yet useless dir just before branching would be counterproductive.  -- 1KB

+++ crawl-ref/source/skills.cc
@@ -695,7 +695,7 @@ static void _check_skill_cost_change()
 void change_skill_points(skill_type sk, int points, bool do_level_up)
 {
     if (static_cast<int>(you.skill_points[sk]) < -points)
-        points = -you.skill_points[sk];
+        points = -(int)you.skill_points[sk];
 
     you.skill_points[sk] += points;
     you.total_skill_points += points;
